+++ src/main/java/com/kap/mechanics_api/dto/movimentacaoestoque/MovimentacaoEstoqueResponseDTO.java
@@ -5,14 +5,14 @@ import com.kap.mechanics_api.enums.TipoMovimentacaoEstoque;
 import java.time.LocalDateTime;
 
 public record MovimentacaoEstoqueResponseDTO(
-        Long id,
+        Integer id,
         Integer itemEstoqueId,
         String itemEstoqueNome,
         TipoMovimentacaoEstoque tipo,
         Integer quantidade,
         LocalDateTime dataHora,
         Integer usuarioId,
-        Long ordemServicoId,
+        Integer ordemServicoId,
         Integer saldoItemEstoque
 ) {
 }
