+++ src/main/java/com/kap/mechanics_api/repository/OrdemServicoRepository.java
@@ -3,7 +3,7 @@ package com.kap.mechanics_api.repository;
 import com.kap.mechanics_api.domain.OrdemServico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdemServicoRepository extends JpaRepository<OrdemServico, Long> {
+public interface OrdemServicoRepository extends JpaRepository<OrdemServico, Integer> {
 
-    boolean existsByOrcamentoId(Long orcamentoId);
+    boolean existsByOrcamentoId(Integer orcamentoId);
 }
